Kevin/Randy - eslint override, import changes,
