+++ frontend/package.json
@@ -287,6 +287,16 @@
                 }
             ]
         },
+        "overrides": [
+            {
+                "files": [
+                    "**/*.test.ts?"
+                ],
+                "rules": {
+                    "@typescript-eslint/no-restricted-imports": "off"
+                }
+            }
+        ],
         "settings": {
             "react": {
                 "version": "detect"
