+++ frontend/src/App.tsx
@@ -52,7 +52,7 @@ import './App.css'
 import ACMPerspectiveIcon from './assets/ACM-icon.svg'
 import logo from './assets/RHACM-Logo.svg?url'
 // eslint-disable-next-line @typescript-eslint/no-restricted-imports
-import { logout } from './atoms'
+import { logout } from './logout'
 import { LoadingPage } from './components/LoadingPage'
 import { getApplinks, IAppSwitcherData } from './lib/applinks'
 import { configure } from './lib/configure'
