+++ frontend/src/atoms.tsx
@@ -1,5 +1,4 @@
 /* Copyright Contributors to the Open Cluster Management project */
-import { noop } from 'lodash'
 import {
     AgentClusterInstallK8sResource,
     AgentK8sResource,
@@ -78,7 +77,6 @@ import {
     DiscoveryConfig,
     DiscoveryConfigApiVersion,
     DiscoveryConfigKind,
-    fetchGet,
     getBackendUrl,
     GitOpsCluster,
     GitOpsClusterApiVersion,
@@ -168,6 +166,7 @@ import {
     UserPreferenceApiVersion,
     UserPreferenceKind,
 } from './resources'
+import { tokenExpired } from './logout'
 let atomArrayKey = 0
 function AtomArray<T>() {
     return atom<T[]>({ key: (++atomArrayKey).toString(), default: [] })
@@ -598,35 +597,4 @@ export function usePolicies() {
 export function useSavedSearchLimit() {
     const settings = useRecoilValue(settingsState)
     return useMemo(() => parseInt(settings.SAVED_SEARCH_LIMIT ?? '10'), [settings])
-}
-
-export async function tokenExpired() {
-    if (process.env.NODE_ENV === 'production') {
-        logout()
-    } else {
-        window.location.href = `${getBackendUrl()}/login`
-    }
-}
-
-export async function logout() {
-    const tokenEndpointResult = await fetchGet<{ token_endpoint: string }>(getBackendUrl() + '/configure')
-    await fetchGet(getBackendUrl() + '/logout').catch(noop)
-
-    const iframe = document.createElement('iframe')
-    iframe.setAttribute('type', 'hidden')
-    iframe.name = 'hidden-form'
-    document.body.appendChild(iframe)
-
-    const form = document.createElement('form')
-    form.method = 'POST'
-    form.target = 'hidden-form'
-    const url = new URL(tokenEndpointResult.data.token_endpoint)
-    form.action = `${url.protocol}//${url.host}/logout`
-    document.body.appendChild(form)
-
-    form.submit()
-
-    await new Promise((resolve) => setTimeout(resolve, 500))
-
-    location.pathname = '/'
 }
