+++ frontend/src/routes/Infrastructure/Clusters/ManagedClusters/CreateCluster/components/assisted-installer/utils.ts
@@ -1,6 +1,6 @@
 /* Copyright Contributors to the Open Cluster Management project */
 /* eslint-disable react-hooks/exhaustive-deps */
-import { useCallback, useMemo, useState, useEffect, useContext } from 'react'
+import { useCallback, useMemo, useState, useEffect } from 'react'
 import { isEqual } from 'lodash'
 import { CIM } from 'openshift-assisted-ui-lib'
